r/aws_datasync_task: Use 'tfresource.SetLastError'.
Acceptance test output:

% make testacc TESTARGS='-run=TestAccDataSyncTask_basic' PKG=datasync ACCTEST_PARALLELISM=3
==> Checking that code complies with gofmt requirements...
TF_ACC=1 go test ./internal/service/datasync/... -v -count 1 -parallel 3  -run=TestAccDataSyncTask_basic -timeout 180m
=== RUN   TestAccDataSyncTask_basic
=== PAUSE TestAccDataSyncTask_basic
=== CONT  TestAccDataSyncTask_basic
--- PASS: TestAccDataSyncTask_basic (206.88s)
PASS
ok  	github.com/hashicorp/terraform-provider-aws/internal/service/datasync	212.840s
